+++ src/main/kotlin/icu/windea/pls/config/config/delegated/CwtSubtypeConfig.kt
@@ -15,9 +15,6 @@ import icu.windea.pls.model.*
  * @property typeKeyFilter (option*) type_key_filter: string | string[]
  * @property typeKeyRegex (option) type_key_regex: string
  * @property startsWith (option) starts_with: string
- * @property pushScope (option) push_scope: scope
- * @property displayName (option) display_name: string
- * @property abbreviation (option) abbreviation: string
  * @property onlyIfNot (option) only_if_not: string[]
  */
 interface CwtSubtypeConfig : CwtDelegatedConfig<CwtProperty, CwtPropertyConfig> {
@@ -25,9 +22,6 @@ interface CwtSubtypeConfig : CwtDelegatedConfig<CwtProperty, CwtPropertyConfig> 
     val typeKeyFilter: ReversibleValue<Set<@CaseInsensitive String>>?
     val typeKeyRegex: Regex?
     val startsWith: @CaseInsensitive String?
-    val pushScope: String?
-    val displayName: String?
-    val abbreviation: String?
     val onlyIfNot: Set<String>?
 
     companion object {
@@ -45,10 +39,7 @@ private fun doResolve(config: CwtPropertyConfig): CwtSubtypeConfig? {
     val name = config.key.removeSurroundingOrNull("subtype[", "]")?.orNull()?.intern() ?: return null
     var typeKeyFilter: ReversibleValue<Set<String>>? = null
     var typeKeyRegex: Regex? = null
-    var pushScope: String? = null
     var startsWith: String? = null
-    var displayName: String? = null
-    var abbreviation: String? = null
     var onlyIfNot: Set<String>? = null
 
     val options = config.optionConfigs.orEmpty()
@@ -69,13 +60,10 @@ private fun doResolve(config: CwtPropertyConfig): CwtSubtypeConfig? {
                 typeKeyRegex = option.stringValue?.toRegex(RegexOption.IGNORE_CASE)
             }
             "starts_with" -> startsWith = option.stringValue ?: continue //忽略大小写
-            "push_scope" -> pushScope = option.stringValue ?: continue
-            "display_name" -> displayName = option.stringValue ?: continue
-            "abbreviation" -> abbreviation = option.stringValue ?: continue
             "only_if_not" -> onlyIfNot = option.getOptionValueOrValues() ?: continue
         }
     }
-    return CwtSubtypeConfigImpl(config, name, typeKeyFilter, typeKeyRegex, startsWith, pushScope, displayName, abbreviation, onlyIfNot?.optimized())
+    return CwtSubtypeConfigImpl(config, name, typeKeyFilter, typeKeyRegex, startsWith, onlyIfNot?.optimized())
 }
 
 private class CwtSubtypeConfigImpl(
@@ -84,9 +72,6 @@ private class CwtSubtypeConfigImpl(
     override val typeKeyFilter: ReversibleValue<Set<String>>? = null,
     override val typeKeyRegex: Regex? = null,
     override val startsWith: String? = null,
-    override val pushScope: String? = null,
-    override val displayName: String? = null,
-    override val abbreviation: String? = null,
     override val onlyIfNot: Set<String>? = null
 ) : UserDataHolderBase(), CwtSubtypeConfig {
     override fun toString(): String {