+++ src/main/kotlin/icu/windea/pls/cwt/codeInsight/completion/CwtConfigCompletionProvider.kt
@@ -12,16 +12,18 @@ class CwtConfigCompletionProvider : CompletionProvider<CompletionParameters>() {
     override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
         val position = parameters.position
         val contextElement: PsiElement? = when (position.elementType) {
+            CwtElementTypes.OPTION_KEY_TOKEN -> position.parent?.castOrNull<CwtOptionKey>()
             CwtElementTypes.PROPERTY_KEY_TOKEN -> position.parent?.castOrNull<CwtPropertyKey>()
-            CwtElementTypes.STRING_TOKEN -> position.parent?.castOrNull<CwtString>()?.takeIf { it.isPropertyValue() || it.isBlockValue() }
+            CwtElementTypes.STRING_TOKEN -> position.parent?.castOrNull<CwtString>()
             else -> null
         }
         if (contextElement == null) return
 
-        val r = CwtConfigCompletionManager.initializeContext(parameters, context, contextElement)
+        val r = CwtConfigCompletionManager.initializeContext(contextElement, parameters, context)
         if (!r) return
+        val r1 = CwtConfigCompletionManager.initializeContextForConfigCompletions(context)
+        if (!r1) return
 
         CwtConfigCompletionManager.addConfigCompletions(context, result)
     }
 }
-