+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 * [X] 检查：预定义的动态模版，相关功能是否正常
 * [ ] 检查：复杂表达式的代码高亮以及颜色主题设置，相关功能是否正常
+* [ ] 检查：对规则文件的代码补全，相关功能是否正常
 
 ## 1.4.1
 
@@ -23,6 +24,7 @@
 * [X] #144 cwt规则的type_key_filter <> xxx失效
 * [X] 修复：优化`PlsFileListener`，尝试避免某些特殊情况下的 SOF / FIX: Optimize `PlsFileListener` to avoid SOF in some special cases
 * [X] 修复：修复封装变量引用的代码折叠规则失效的问题 / FIX: Fixed the problem that the code folding rule for scripted variable references is inactive
+* [X] 新功能：初步提供对规则文件中选项的代码补全 / NEW: Basic support for code completion of options in config files *
 * [X] 其他优化与BUG修复 / Other optimizations and bug fixes
 
 ## 1.4.0